# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "投资收益统计系统 API",
    "description": "方案B — 资产快照 + 出入金 + 证券买卖 + 标的最新价 + 现金余额",
    "version": "1.0"
  },
  "paths": {
...
    "/api/portfolios": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolios",
        "operationId": "list_portfolios_api_portfolios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioOut"
                  },
                  "type": "array",
                  "title": "Response List Portfolios Api Portfolios Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio",
        "operationId": "create_portfolio_api_portfolios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreateReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/api/portfolios/{portfolio_id}/export": {
      "get": {
        "tags": [
          "data-transfer"
        ],
        "summary": "Export Data",
        "operationId": "export_data_api_portfolios__portfolio_id__export_get",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "securities",
              "title": "Type"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "csv",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/import/preview": {
      "post": {
        "tags": [
          "data-transfer"
        ],
        "summary": "Import Preview",
        "operationId": "import_preview_api_portfolios__portfolio_id__import_preview_post",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_preview_api_portfolios__portfolio_id__import_preview_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/import/commit": {
      "post": {
        "tags": [
          "data-transfer"
        ],
        "summary": "Import Commit",
        "operationId": "import_commit_api_portfolios__portfolio_id__import_commit_post",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportCommitReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportCommitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data-transfer/template": {
      "get": {
        "tags": [
          "data-transfer"
        ],
        "summary": "Download Template",
        "operationId": "download_template_api_data_transfer_template_get",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "securityTrades",
              "title": "Type"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "csv",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_preview_api_portfolios__portfolio_id__import_preview_post": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "type",
          "file"
        ],
        "title": "Body_import_preview_api_portfolios__portfolio_id__import_preview_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportCommitOut": {
        "properties": {
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "failed": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Failed",
            "default": []
          },
          "recalculated": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recalculated"
          }
        },
        "type": "object",
        "required": [
          "inserted",
          "updated",
          "skipped"
        ],
        "title": "ImportCommitOut"
      },
      "ImportCommitReq": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "type",
          "token"
        ],
        "title": "ImportCommitReq"
      },
      "ImportPreviewOut": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "totalRows": {
            "type": "integer",
            "title": "Totalrows"
          },
          "validRows": {
            "type": "integer",
            "title": "Validrows"
          },
          "sample": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sample",
            "default": []
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "minDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mindate"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "type",
          "totalRows",
          "validRows",
          "token"
        ],
        "title": "ImportPreviewOut"
      },
...
      "PortfolioCreateReq": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PortfolioCreateReq"
      },
      "PortfolioOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "baseDate": {
            "type": "string",
            "format": "date",
            "title": "Basedate"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "archivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "baseDate",
          "currency",
          "createdAt",
          "updatedAt"
        ],
        "title": "PortfolioOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "JWT-auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}